start to migrate toward standard phpunit framework
pretty much the entire phpunit framework is broken right now, throwing
errors and eventually running out of memory.  PHPUnit 3.4 is more
flexible than 1.x was, so I'm not sure that we need the custom test
runner anymore.  I'm not sure how we'll support the custom cli options
like --no-math...  perhaps we'll figure that out, but for now it's more
important to get things working again, and using the standard phpunit
framework seems to be the fastest and most stable way to do that.